chnages 2026

+++ src/app/core/services/account.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
 import { Account } from '../models/account.model';
 import { environment } from '../../../environments/environment';
@@ -16,7 +17,11 @@ export class AccountService {
   private accountCache$?: Observable<Account> | null;
   private logoutTimer?: any;
 
-  constructor(private http: HttpClient, private router: Router) {}
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) { }
 
   getUserDetails(): Observable<Account> {
     return this.http.get<Account>(`${this.baseUrl}/auth/users/me/`);
@@ -26,23 +31,25 @@ export class AccountService {
     if (!this.accountCache$ || force) {
       this.accountCache$ = this.getUserDetails().pipe(
         tap(account => {
-          localStorage.setItem('username', account?.username ?? '');
-          localStorage.setItem('role', account.role!.name);
-          localStorage.setItem('firstName', account.firstName);
-          localStorage.setItem('lastName', account.lastName);
+          if (isPlatformBrowser(this.platformId)) {
+            localStorage.setItem('username', account?.username ?? '');
+            localStorage.setItem('role', account.role!.name);
+            localStorage.setItem('firstName', account.firstName);
+            localStorage.setItem('lastName', account.lastName);
+
+            // Auto logout
+            const access = localStorage.getItem('access');
+            const expiry = access ? TokenUtil.getTokenExpiry(access) : null;
+            if (expiry) {
+              const timeout = expiry - Date.now();
+              this.logoutTimer = setTimeout(() => {
+                this.clearAppData();
+                this.router.navigate(['/login'], { queryParams: { sessionExpired: true } });
+              }, timeout);
+            }
+          }
 
           this.authenticate(account);
-
-          // Auto logout
-          const access = localStorage.getItem('access');
-          const expiry = access ? TokenUtil.getTokenExpiry(access) : null;
-          if (expiry) {
-            const timeout = expiry - Date.now();
-            this.logoutTimer = setTimeout(() => {
-              this.clearAppData();
-              this.router.navigate(['/login'], { queryParams: { sessionExpired: true } });
-            }, timeout);
-          }
         }),
         shareReplay()
       );
@@ -82,8 +89,10 @@ export class AccountService {
       this.logoutTimer = undefined;
     }
 
-    localStorage.clear();
-    sessionStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+      sessionStorage.clear();
+    }
     this.authenticate(null);
   }
 }
